refactor: TakeSnapshot example to TS and FC
